+++ packages/formatjs/packages/cli/tests/extract/integration_tests/defineMessages/actual.js
@@ -27,6 +27,14 @@ const msgs = defineMessages({
     description: 'Escaped apostrophe',
     defaultMessage: "A quoted value ''{value}'",
   },
+  noId: {
+    description: 'no ID',
+    defaultMessage: "No ID",
+  },
+  duplicateAsNoId: {
+    description: 'no ID',
+    defaultMessage: "No ID",
+  }
 });
 
 export default class Foo extends Component {