+++ packages/formatjs/packages/cli/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.0.1...@formatjs/cli@1.0.2) (2019-10-31)
+
+
+### Bug Fixes
+
+* **@formatjs/cli:** Add support for parsing class properties ([1714bf9](https://github.com/formatjs/formatjs/commit/1714bf9d9abfc31458222e0922ef57baa1f38787))
+
+
+
+
+
 ## [1.0.1](https://github.com/formatjs/formatjs/compare/@formatjs/cli@1.0.0...@formatjs/cli@1.0.1) (2019-10-31)